+++ ethereum/graphql/src/test/resources/tech/pegasys/pantheon/ethereum/graphql/eth_getBalance_invalidAccountLatest.json
@@ -1,7 +1,6 @@
 {
   "request": "{account(address: \"0xdeaff00ddeaff00ddeaff00ddeaff00ddeaff00d\") { balance } }",
   "response": {
-    "data": null,
     "errors": [
       {
         "message": "Exception while fetching data (/account) : Account with address 0xdeaff00ddeaff00ddeaff00ddeaff00ddeaff00d does not exist",
@@ -13,9 +12,13 @@
         ],
         "path": [
           "account"
-        ]
+        ],
+        "extensions": {
+          "classification": "DataFetchingException"
+        }
       }
-    ]
+    ],
+    "data": null
   },
   "statusCode": 400
 }
