+++ ethereum/graphql/src/test/resources/tech/pegasys/pantheon/ethereum/graphql/eth_getBalance_without_addr.json
@@ -9,7 +9,10 @@
             "line": 1,
             "column": 2
           }
-        ]
+        ],
+        "extensions": {
+          "classification": "ValidationError"
+        }
       }
     ]
   },
