+++ ethereum/graphql/src/test/resources/tech/pegasys/pantheon/ethereum/graphql/eth_getBlock_byNumberInvalid.json
@@ -1,7 +1,6 @@
 {
   "request": "{block (number: 88888888) {number }} ",
   "response": {
-    "data": null,
     "errors": [
       {
         "message": "Exception while fetching data (/block) : Block number 88888888 was not found",
@@ -13,9 +12,13 @@
         ],
         "path": [
           "block"
-        ]
+        ],
+        "extensions": {
+          "classification": "DataFetchingException"
+        }
       }
-    ]
+    ],
+    "data": null
   },
   "statusCode": 400
 }
