+++ ethereum/graphql/src/test/resources/tech/pegasys/pantheon/ethereum/graphql/eth_getBlock_wrongParams.json
@@ -1,7 +1,6 @@
 {
   "request": "{block (number: \"0x03\", hash : \"0xc8df1f061abb4d0c107b2b1a794ade8780b3120e681f723fe55a7be586d95ba6\") {number transactions{hash} timestamp difficulty totalDifficulty gasUsed gasLimit hash nonce ommerCount logsBloom mixHash ommerHash extraData stateRoot receiptsRoot transactionCount transactionsRoot}} ",
   "response": {
-    "data": null,
     "errors": [
       {
         "message": "Exception while fetching data (/block) : Invalid params",
@@ -16,10 +15,12 @@
         ],
         "extensions": {
           "errorCode": -32602,
-          "errorMessage": "Invalid params"
+          "errorMessage": "Invalid params",
+          "classification": "DataFetchingException"
         }
       }
-    ]
+    ],
+    "data": null
   },
   "statusCode": 400
 }
