+++ ethereum/graphql/src/test/resources/tech/pegasys/pantheon/ethereum/graphql/eth_sendRawTransaction_nonceTooLow.json
@@ -15,7 +15,8 @@
         ],
         "extensions": {
           "errorCode": -32001,
-          "errorMessage": "Nonce too low"
+          "errorMessage": "Nonce too low",
+          "classification": "DataFetchingException"
         }
       }
     ],
