+++ ethereum/graphql/src/test/resources/tech/pegasys/pantheon/ethereum/graphql/eth_sendRawTransaction_unsignedTransaction.json
@@ -1,19 +1,26 @@
 {
-  "request" : "mutation { sendRawTransaction(data: \"0xed0a85174876e800830222e0945aae326516b4f8fe08074b7e972e40a713048d62880de0b6b3a7640000801c8080\") }",
-  "response":{
-    "data" : null,
-    "errors" : [ {
-      "message" : "Exception while fetching data (/sendRawTransaction) : Invalid params",
-      "locations" : [ {
-        "line" : 1,
-        "column" : 12
-      } ],
-      "path" : [ "sendRawTransaction" ],
-      "extensions" : {
-        "errorCode" : -32602,
-        "errorMessage" : "Invalid params"
+  "request": "mutation { sendRawTransaction(data: \"0xed0a85174876e800830222e0945aae326516b4f8fe08074b7e972e40a713048d62880de0b6b3a7640000801c8080\") }",
+  "response": {
+    "errors": [
+      {
+        "message": "Exception while fetching data (/sendRawTransaction) : Invalid params",
+        "locations": [
+          {
+            "line": 1,
+            "column": 12
+          }
+        ],
+        "path": [
+          "sendRawTransaction"
+        ],
+        "extensions": {
+          "errorCode": -32602,
+          "errorMessage": "Invalid params",
+          "classification": "DataFetchingException"
+        }
       }
-    } ]
+    ],
+    "data": null
   },
   "statusCode": 400
 }
